GEMM reads its micro-kernel geometry from the kernel factory
The class carried two copies of (r, s, t) and they disagreed about what
they were for. mac_dims() went through a microkernel_mac_dim_map tabulated
here; validate(), ten lines below it, spelled out (8, 8, 8) / (4, 8, 8)
again. Both now call aie.iron.kernels.mm_mac_dims, which is the table
upstream keeps in step with mm.cc's own combos(X) X(..., r, s, t) macros.

validate() keeps asking for aie2p, which is what it has always assumed and
what its messages name, since the source it compiles is
aie_kernels/aie2p/mm.cc; there is no device at construction to ask anyway.
design() asks for the device it is building for, so npu1 still gets the
looser (4, 8, 4). Those two have always differed; naming the reason is new.

Also drops the dtype=data.dtype the harness needed while NpuTensor's
constructor ignored a typed array's dtype (mlir-aie 20e6836).

+++ iron/common/harness.py
@@ -205,9 +205,7 @@ def run_test(
                 produced[name] = buf
             else:
                 name, data = next(ins)
-                # dtype explicitly: the constructor's default is uint32, so a
-                # typed array would otherwise be reinterpreted, not adopted.
-                buf = tensor_class(data, dtype=data.dtype)
+                buf = tensor_class(data)
                 if b.direction == "inout":
                     produced[name] = buf
         except StopIteration:
